かかと落とし実装と天使の攻撃を追加

+++ SummerGame/Game/Enemy.cpp
@@ -96,7 +96,7 @@ void Enemy::Update()
 	if (m_attackCooldown > 0)
 	{
 		//攻撃のクールタイムもウィッチタイムで遅くする
-		m_attackCooldown -= scale;
+		m_attackCooldown -= static_cast<int>(scale);
 	}
 
 	//ステート
